bookings status style fix
